+++ example/src/app/(home)/index.tsx
@@ -36,7 +36,7 @@ const cards: HomeCardProps[] = [
       'https://heroui-assets.nyc3.cdn.digitaloceanspaces.com/images/heroui-native-example/home-components-light.png',
     imageDark:
       'https://heroui-assets.nyc3.cdn.digitaloceanspaces.com/images/heroui-native-example/home-components-dark.png',
-    count: 20,
+    count: 21,
     footer: 'Explore all components',
     path: 'components',
   },
@@ -149,7 +149,9 @@ export default function App() {
   return (
     <ScreenScrollView>
       <View className="items-center justify-center my-4">
-        <AppText className="text-muted text-base">v1.0.0-alpha.15</AppText>
+        <AppText className="text-muted-foreground text-base">
+          v1.0.0-alpha.16
+        </AppText>
       </View>
       <View className="gap-6">
         {cards.map((card, index) => (